@@ -1,18 +1,16 @@
-const express = require("express");
+import express from "express";
+import * as Brewery from "../controllers/brewery";
 
 const router = express.Router();
-const Brewery = require("../controllers/brewery");
 
 router.get("/brewery", Brewery.getAll);
 router.get("/brewery/:id", Brewery.getById);
-
 // router.get("/brewery/country/:country", (req, res, next) => {
 //   Brewery.find({ country: req.params.country })
 //     .then((data) => res.json(data))
 //     .catch(next);
 // });
-
 router.post("/brewery", Brewery.add);
 router.delete("/brewery/:id", Brewery.remove);
 
-module.exports = router;
+export default router;
